Initial rewrite

+++ environment/sql/tables/metric.sql
@@ -3,8 +3,8 @@ CREATE TABLE `metric` (
   `name` varchar(30) NOT NULL,
   `date` TIMESTAMP NOT NULL,
   `commandFK` int(11) NULL,
-  `guildFK` bigint(11) NULL,
-  `userFK` bigint(11) NULL,
+  `guildFK` int(11) NULL,
+  `userFK` int(11) NULL,
   `value` bigint(11) NULL,
   `details` varchar(200) NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
