+++ environment/sql/tables/role.sql
@@ -1,6 +1,6 @@
 CREATE TABLE `role` (
   `id` int(11) PRIMARY KEY AUTO_INCREMENT,
-  `guildFK` bigint(11) NOT NULL,
+  `guildFK` int(11) NOT NULL,
   `parent` int(11) DEFAULT NULL,
   `name` varchar(30) NOT NULL,
   `description` varchar(500) NOT NULL
